% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predisposition_gene.R
\name{append_cpg_properties}
\alias{append_cpg_properties}
\title{Function that appends mechanism-of-inheritance (MOI) and mechanism
of disease (MOD) annotations to cancer predisposition genes (cpg's),
as well as estimated fractions of truncation vs. non-truncating variants
etc per predisposition gene (from ClinVar)}
\usage{
append_cpg_properties(cpg_calls, ref_data = NULL)
}
\arguments{
\item{cpg_calls}{data frame with variant calls in predisposition genes}

\item{ref_data}{PCGR/CPSR reference data object}
}
\value{
cpg_calls data frame with cancer predisposition gene properties
appended (mechanism of disease, inheritance patterns etc)
}
\description{
Function that appends mechanism-of-inheritance (MOI) and mechanism
of disease (MOD) annotations to cancer predisposition genes (cpg's),
as well as estimated fractions of truncation vs. non-truncating variants
etc per predisposition gene (from ClinVar)
}
