% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_acmg_consensus}
\alias{assign_acmg_consensus}
\title{Function that assigns a summary string of
all ACMG evidence codes met for a given variant
(e.g. "PVS1|PS1|PM2_supporting|PP3")}
\usage{
assign_acmg_consensus(var_calls = NULL)
}
\arguments{
\item{var_calls}{data frame with variant calls in predisposition genes}
}
\value{
var_calls data frame with ACMG_CODE column appended
}
\description{
Function that assigns a summary string of
all ACMG evidence codes met for a given variant
(e.g. "PVS1|PS1|PM2_supporting|PP3")
}
