% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_acmg_evidence}
\alias{assign_acmg_evidence}
\title{Function that assigns final pathogenicity classification (B, LB, VUS, P, LP)
based on accumulated scores from different ACMG criteria and pre-defined
cutoffs (calibrated against ClinVar)}
\usage{
assign_acmg_evidence(var_calls, settings, ref_data)
}
\arguments{
\item{var_calls}{sample calls with dbnsfp annotations}

\item{settings}{cpsr settings object}

\item{ref_data}{pcgr data object}
}
\value{
calls
}
\description{
param var_calls data frame with variant calls in predisposition genes
}
\details{
return var_calls data frame with pathogenicity classification appended

export
Function that assigns variant pathogenicity evidence based on ACMG guidelines
}
