% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_BS1_evidence}
\alias{assign_BS1_evidence}
\title{Function that assigns ACMG evidence indicators for
BS1_ST and BS1_SUP}
\usage{
assign_BS1_evidence(var_calls = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}
}
\value{
data.frame with ACMG evidence indicators
}
\description{
Function that assigns ACMG evidence indicators for
BS1_ST and BS1_SUP
}
