% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_classification}
\alias{assign_classification}
\title{Function that assigns final pathogenicity classification (B, LB, VUS, P, LP)
based on accumulated scores from different ACMG criteria and pre-defined
cutoffs (calibrated against ClinVar)}
\usage{
assign_classification(var_calls)
}
\arguments{
\item{var_calls}{data frame with variant calls in predisposition genes}
}
\value{
var_calls data frame with pathogenicity classification appended
}
\description{
Function that assigns final pathogenicity classification (B, LB, VUS, P, LP)
based on accumulated scores from different ACMG criteria and pre-defined
cutoffs (calibrated against ClinVar)
}
