% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_classification_authority}
\alias{assign_classification_authority}
\title{Function that combines classifications of novel and
pre-classified variants}
\usage{
assign_classification_authority(var_calls, conf = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}

\item{conf}{CPSR configuration object with run settings}
}
\description{
Function that combines classifications of novel and
pre-classified variants
}
