% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_PM1_evidence}
\alias{assign_PM1_evidence}
\title{Function that assigns ACMG evidence indicators for
PM1 ('ACMG_PM1' and 'ACMG_PM1_SUPP')}
\usage{
assign_PM1_evidence(var_calls = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}
}
\value{
data.frame with ACMG evidence indicators
}
\description{
Function that assigns ACMG evidence indicators for
PM1 ('ACMG_PM1' and 'ACMG_PM1_SUPP')
}
