% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_PM2_evidence}
\alias{assign_PM2_evidence}
\title{Function that assigns ACMG evidence indicators for
PM2_supporting}
\usage{
assign_PM2_evidence(var_calls = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}
}
\value{
data.frame with ACMG evidence indicators
}
\description{
Function that assigns ACMG evidence indicators for
PM2_supporting
}
