% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_PM5_evidence}
\alias{assign_PM5_evidence}
\title{Function that assigns ACMG evidence indicators for
PM5}
\usage{
assign_PM5_evidence(var_calls = NULL, pathogenic_codons = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}

\item{pathogenic_codons}{data.frame with codons where
pathogenic variants have been observed (from ClinVar)}
}
\value{
data.frame with ACMG evidence indicators
}
\description{
Function that assigns ACMG evidence indicators for
PM5
}
