% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_PP2_evidence}
\alias{assign_PP2_evidence}
\title{Function that assigns ACMG evidence indicators for
PP3}
\usage{
assign_PP2_evidence(
  var_calls = NULL,
  max_benign_missense_frac = 0.1,
  max_truncating_frac = 0.5
)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}
}
\value{
data.frame with ACMG evidence indicators
}
\description{
Function that assigns ACMG evidence indicators for
PP3
}
