% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_PP3_evidence}
\alias{assign_PP3_evidence}
\title{Function that assigns ACMG evidence indicators for
PP3}
\usage{
assign_PP3_evidence(var_calls = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}
}
\value{
data.frame with ACMG evidence indicators
}
\description{
Function that assigns ACMG evidence indicators for
PP3
}
