% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_PS1_evidence}
\alias{assign_PS1_evidence}
\title{Function that assigns ACMG evidence indicators for
PS1 ('ACMG_PS1')}
\usage{
assign_PS1_evidence(
  var_calls = NULL,
  pathogenic_codons = NULL,
  pathogenic_nucleotides = NULL
)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}

\item{pathogenic_codons}{data.frame with known pathogenic codons
pathogenic variants have been observed (from ClinVar)}

\item{pathogenic_nucleotides}{data.frame with pathogenic
nucleotides (non-coding, e.g. splice)}
}
\value{
data.frame with ACMG evidence indicators
}
\description{
Function that assigns ACMG evidence indicators for
PS1 ('ACMG_PS1')
}
