% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_PVS1_evidence}
\alias{assign_PVS1_evidence}
\title{Function that assigns ACMG evidence indicators for
PVS1 ('ACMG_PVS1', 'ACMG_PVS1_STR', 'ACMG_PVS1_MOD')}
\usage{
assign_PVS1_evidence(var_calls = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}
}
\value{
data.frame with ACMG evidence indicators
}
\description{
Function that assigns ACMG evidence indicators for
PVS1 ('ACMG_PVS1', 'ACMG_PVS1_STR', 'ACMG_PVS1_MOD')
}
