% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsr_functions.R
\name{assign_variant_tiers}
\alias{assign_variant_tiers}
\title{Function that assign variants to different tiers for
prioritization of germline variants}
\usage{
assign_variant_tiers(cpg_calls, config = NULL, col_format_output = NULL)
}
\arguments{
\item{cpg_calls}{data frame with variants in predisposition_genes}

\item{config}{CPSR configuration object with run settings}

\item{col_format_output}{object with output column formats (html, tsv)}
}
\description{
Function that assign variants to different tiers for
prioritization of germline variants
}
