% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx.R
\name{canonical_pgx_key}
\alias{canonical_pgx_key}
\title{Canonical (sorted) representation of a CPIC lookup key}
\usage{
canonical_pgx_key(key_elements)
}
\arguments{
\item{key_elements}{character vector with lookup key elements, e.g.
c('1.0:1', '0.5:1') or c('TPMT:Normal Metabolizer', 'NUDT15:...')}
}
\description{
Canonical (sorted) representation of a CPIC lookup key
}
\keyword{internal}
