% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{clinvar_trust_levels_doc_note}
\alias{clinvar_trust_levels_doc_note}
\title{Get documentation string for the ClinVar trust levels}
\usage{
clinvar_trust_levels_doc_note(trust_level = NULL)
}
\arguments{
\item{trust_level}{trust level used for the report (0-4), marked in the
table; NULL for none}
}
\value{
A documentation string (markdown)
}
\description{
Table of the levels of the 'clinvar_trust_level' setting, i.e. when the
CPSR classification takes precedence over an existing ClinVar
classification (as implemented in assign_classification_authority())
}
