% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_format_output}
\alias{col_format_output}
\title{Format of CPSR output data frames (HTML, TSV)}
\format{
A list object with four elements: 'html_tier', 'html_sf',
'tsv','html_gwas'
}
\usage{
col_format_output
}
\description{
Format of CPSR output data frames (HTML, TSV)
}
\keyword{datasets}
