% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_palette}
\alias{color_palette}
\title{CPSR color palette}
\format{
A list object with three elements: 'report', 'none', 'genotypes'
}
\usage{
color_palette
}
\description{
CPSR color palette
}
\keyword{datasets}
