% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{combine_novel_and_preclassified}
\alias{combine_novel_and_preclassified}
\title{Function that combines classifications of novel and
pre-classified variants}
\usage{
combine_novel_and_preclassified(var_calls, conf = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}

\item{conf}{CPSR configuration object with run settings}
}
\description{
Function that combines classifications of novel and
pre-classified variants
}
