% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{cpsr_intro_doc_note}
\alias{cpsr_intro_doc_note}
\title{Get documentation string for the CPSR report introduction}
\usage{
cpsr_intro_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for the CPSR report introduction
}
