% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{create_variant_table_theme}
\alias{create_variant_table_theme}
\title{Applies consistent styling using CPSR color palette.}
\usage{
create_variant_table_theme(color_palette = NULL, header_color = NULL)
}
\arguments{
\item{color_palette}{CPSR color_palette object}

\item{header_color}{Optional hex color for the header background.
Defaults to color_palette$report.}
}
\value{
reactableTheme object
}
\description{
Applies consistent styling using CPSR color palette.
}
