% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_pdf.R
\name{doc_note_pdf}
\alias{doc_note_pdf}
\title{Convert the HTML elements of a documentation note (markdown with inline
HTML, as used in the HTML report) to markdown, for the PDF report}
\usage{
doc_note_pdf(x)
}
\arguments{
\item{x}{documentation note (character)}
}
\value{
documentation note with \if{html}{\out{<b>}}/\if{html}{\out{<i>}} as markdown emphasis, and line
breaks (\if{html}{\out{<br>}}) removed
}
\description{
Convert the HTML elements of a documentation note (markdown with inline
HTML, as used in the HTML report) to markdown, for the PDF report
}
\keyword{internal}
