% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{exclude_vars_by_maf}
\alias{exclude_vars_by_maf}
\title{Function that filters novel (non-ClinVar) variants based
on a MAF threshold in gnomAD}
\usage{
exclude_vars_by_maf(var_calls = NULL, conf = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}

\item{conf}{CPSR configuration object with run settings}
}
\description{
Function that filters novel (non-ClinVar) variants based
on a MAF threshold in gnomAD
}
