% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{exclude_vars_non_cancer}
\alias{exclude_vars_non_cancer}
\title{Function that ignores/exclude variants attributed to non-cancer
related phenotypes in ClinVar}
\usage{
exclude_vars_non_cancer(var_calls = NULL, conf = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}

\item{conf}{CPSR configuration object with run settings}
}
\description{
Function that ignores/exclude variants attributed to non-cancer
related phenotypes in ClinVar
}
