% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx.R
\name{format_pgx_gene_text}
\alias{format_pgx_gene_text}
\title{Readable gene-keyed CPIC text, e.g. implications ('GENE:text;GENE:text')}
\usage{
format_pgx_gene_text(x)
}
\description{
Readable gene-keyed CPIC text, e.g. implications ('GENE:text;GENE:text')
}
\keyword{internal}
