% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_cpsr_report}
\alias{generate_cpsr_report}
\title{Function that generates variant predisposition report - CPSR}
\usage{
generate_cpsr_report(yaml_fname = NULL)
}
\arguments{
\item{yaml_fname}{YAML file with paths to pre-processed
molecular input files, configuration settings, metadata for
reference data bundle etc.}
}
\description{
Function that generates variant predisposition report - CPSR
}
