% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{genotype_label}
\alias{genotype_label}
\title{Display label of a CPSR genotype code}
\usage{
genotype_label(genotype)
}
\arguments{
\item{genotype}{genotype code(s) ('het', 'hom_alt', 'hom_ref',
'undefined')}
}
\value{
readable genotype(s), e.g. 'heterozygous', 'homozygous'
}
\description{
Display label of a CPSR genotype code
}
\keyword{internal}
