% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{gnomad_germline_doc_note}
\alias{gnomad_germline_doc_note}
\title{Get documentation string for gnomAD usage in germline variant classification}
\usage{
gnomad_germline_doc_note()
}
\value{
A documentation string
}
\description{
Get documentation string for gnomAD usage in germline variant classification
}
