% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{html_link_to_tag}
\alias{html_link_to_tag}
\title{Convert an HTML link string ("\if{html}{\out{<a href='...'>}}text\if{html}{\out{</a>}}") into an
htmltools tag (raw HTML is not rendered in reactable row details)}
\usage{
html_link_to_tag(value)
}
\description{
Convert an HTML link string ("\if{html}{\out{<a href='...'>}}text\if{html}{\out{</a>}}") into an
htmltools tag (raw HTML is not rendered in reactable row details)
}
\keyword{internal}
