% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx.R
\name{infer_pgx_gene_phenotype}
\alias{infer_pgx_gene_phenotype}
\title{Infer CPIC phenotype for a single gene}
\usage{
infer_pgx_gene_phenotype(gene_hits, gene_phenotypes, sex = "UNKNOWN")
}
\arguments{
\item{gene_hits}{allele matches (match_pgx_alleles) for the gene,
altered function alleles only}

\item{gene_phenotypes}{CPIC function/activity value to phenotype
mappings for the gene}

\item{sex}{sample sex ('MALE', 'FEMALE' or 'UNKNOWN')}
}
\description{
Infer CPIC phenotype for a single gene
}
\keyword{internal}
