% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insilico_path_predictors}
\alias{insilico_path_predictors}
\title{Insilico protein impact prediction methods}
\format{
A character vector with names of insilico protein impact prediction methods
used in CPSR (from dbNSFP)
}
\usage{
insilico_path_predictors
}
\description{
Insilico protein impact prediction methods
}
\keyword{datasets}
