% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_clinvar_cancer_phenotype}
\alias{is_clinvar_cancer_phenotype}
\title{Function that retrieves variants in cancer predisposition genes linked
to cancer-related conditions according to ClinVar}
\usage{
is_clinvar_cancer_phenotype(cpg_calls, ref_data)
}
\arguments{
\item{cpg_calls}{data frame with variants in predisposition_genes}

\item{ref_data}{object with PCGR/CPSR reference data}
}
\description{
Function that retrieves variants in cancer predisposition genes linked
to cancer-related conditions according to ClinVar
}
