% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{is_purely_intronic_deletion}
\alias{is_purely_intronic_deletion}
\title{Function that checks if a variant is a purely intronic deletion
(i.e. deletion of intronic sequence that does not extend into exonic sequence
at either end)}
\usage{
is_purely_intronic_deletion(hgvsc)
}
\arguments{
\item{hgvsc}{HGVSc notation for variant}
}
\value{
logical indicating whether variant is a purely intronic deletion
}
\description{
Function that checks if a variant is a purely intronic deletion
(i.e. deletion of intronic sequence that does not extend into exonic sequence
at either end)
}
