% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{load_germline_snv_indel}
\alias{load_germline_snv_indel}
\title{Function that reads and validates an annotated germline SNV/InDel
file from CPSR pre-reporting pipeline}
\usage{
load_germline_snv_indel(fname = NA, ref_data = NULL, settings = NULL)
}
\arguments{
\item{fname}{Path to file name}

\item{ref_data}{Object with PCGR/CPSR reference data}

\item{settings}{Object with CPSR report configuration}
}
\description{
Function that reads and validates an annotated germline SNV/InDel
file from CPSR pre-reporting pipeline
}
