% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{min_distance_to_pathogenic}
\alias{min_distance_to_pathogenic}
\title{Function that calculates the closest distance
to a known pathogenic variant in the same gene}
\usage{
min_distance_to_pathogenic(var_calls = NULL, ref_data = NULL)
}
\arguments{
\item{var_calls}{variants in cancer predisposition genes}

\item{ref_data}{reference data object with ClinVar
annotations}
}
\value{
data.frame with closest distance to known
pathogenic variant in same gene
}
\description{
Function that calculates the closest distance
to a known pathogenic variant in the same gene
}
