% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_pdf.R
\name{pgx_genotype_pill_pdf}
\alias{pgx_genotype_pill_pdf}
\title{Genotypes of the detected variants of a gene ('; '-separated:
heterozygous, homozygous, hemizygous) as colored pills in a PDF report
table, one per line - colors as in the HTML report}
\usage{
pgx_genotype_pill_pdf(genotypes)
}
\arguments{
\item{genotypes}{genotypes per table row}
}
\value{
character vector with Typst markup (see pill_pdf)
}
\description{
Genotypes of the detected variants of a gene ('; '-separated:
heterozygous, homozygous, hemizygous) as colored pills in a PDF report
table, one per line - colors as in the HTML report
}
\keyword{internal}
