% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{pgx_phenotype_color}
\alias{pgx_phenotype_color}
\title{Color of a CPIC phenotype - blues, darker = more impact (HTML and PDF)}
\usage{
pgx_phenotype_color(phenotype)
}
\arguments{
\item{phenotype}{CPIC phenotype(s)}
}
\value{
hex color(s); grey for indeterminate/not determined
}
\description{
Color of a CPIC phenotype - blues, darker = more impact (HTML and PDF)
}
\keyword{internal}
