% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_pdf.R
\name{pgx_phenotype_pill_pdf}
\alias{pgx_phenotype_pill_pdf}
\title{CPIC phenotypes as colored pills in a PDF report table - blues, darker
for more impact (pgx_phenotype_color, as in the HTML report)}
\usage{
pgx_phenotype_pill_pdf(phenotypes)
}
\arguments{
\item{phenotypes}{CPIC phenotype per table row}
}
\value{
character vector with Typst markup (see pill_pdf)
}
\description{
CPIC phenotypes as colored pills in a PDF report table - blues, darker
for more impact (pgx_phenotype_color, as in the HTML report)
}
\keyword{internal}
