% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{pgx_phenotype_row_details}
\alias{pgx_phenotype_row_details}
\title{Row details (variant level) for the CPIC phenotype table}
\usage{
pgx_phenotype_row_details(phenotypes, variants = NULL)
}
\arguments{
\item{phenotypes}{data frame with gene phenotypes}

\item{variants}{data frame with pharmacogenomic variants}
}
\value{
function(index) for reactable details
}
\description{
Row details (variant level) for the CPIC phenotype table
}
\keyword{internal}
