% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{pgx_recommendation_row_details}
\alias{pgx_recommendation_row_details}
\title{Row details for the CPIC prescribing recommendation table}
\usage{
pgx_recommendation_row_details(recommendations)
}
\arguments{
\item{recommendations}{data frame with drug recommendations}
}
\value{
function(index) for reactable details
}
\description{
Row details for the CPIC prescribing recommendation table
}
\keyword{internal}
