% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_virtual_panels}
\alias{plot_virtual_panels}
\title{Function that makes a HTML display of virtual gene panel}
\usage{
plot_virtual_panels(gene_df)
}
\arguments{
\item{gene_df}{data frame with genes targeted in virtual panel}
}
\description{
Function that makes a HTML display of virtual gene panel
}
