% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{prep_biomarker_tbl}
\alias{prep_biomarker_tbl}
\title{Function that gathers data table on biomarker variants
for display in germline report}
\usage{
prep_biomarker_tbl(rep = NULL, variant_category = "snv_indel")
}
\arguments{
\item{rep}{report object}

\item{variant_category}{variant category}
}
\description{
Function that gathers data table on biomarker variants
for display in germline report
}
