% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx.R
\name{retrieve_pgx_calls_clinvar}
\alias{retrieve_pgx_calls_clinvar}
\title{Function that retrieves pharmacogenomic variants based on ClinVar
classifications (legacy approach, used for data bundles without CPIC
allele definitions)}
\usage{
retrieve_pgx_calls_clinvar(calls)
}
\arguments{
\item{calls}{data frame with all calls found}
}
\description{
Function that retrieves pharmacogenomic variants based on ClinVar
classifications (legacy approach, used for data bundles without CPIC
allele definitions)
}
\keyword{internal}
