% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{rt_cell_bm_significance}
\alias{rt_cell_bm_significance}
\title{Cell renderer factory for biomarker clinical significance
Returns a function that renders biomarker clinical significance
values as colored pills based on the provided color palette.}
\usage{
rt_cell_bm_significance(color_palette)
}
\arguments{
\item{color_palette}{CPSR color palette object containing
biomarker_types styling information}
}
\value{
A function that takes a biomarker clinical significance value and returns an HTML div element containing styled pills for each significance category
}
\description{
Cell renderer factory for biomarker clinical significance
Returns a function that renders biomarker clinical significance
values as colored pills based on the provided color palette.
}
