% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{rt_cell_pgx_genotype}
\alias{rt_cell_pgx_genotype}
\title{Cell renderer factory for the genotypes of the detected variants of a
gene ('; '-separated: heterozygous, homozygous, hemizygous) - pills
styled as the genotypes of the variant classification table}
\usage{
rt_cell_pgx_genotype(color_palette)
}
\arguments{
\item{color_palette}{CPSR color palette object}
}
\description{
Cell renderer factory for the genotypes of the detected variants of a
gene ('; '-separated: heterozygous, homozygous, hemizygous) - pills
styled as the genotypes of the variant classification table
}
\keyword{internal}
