% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{rt_cell_pgx_phenotype}
\alias{rt_cell_pgx_phenotype}
\title{Cell renderer for CPIC phenotypes - blue badge, darker = more impact}
\usage{
rt_cell_pgx_phenotype(value)
}
\description{
Cell renderer for CPIC phenotypes - blue badge, darker = more impact
}
\keyword{internal}
