% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{rt_cell_pgx_variants}
\alias{rt_cell_pgx_variants}
\title{Cell renderer for the detected variants of a gene ('; '-separated),
one variant per line (aligned with rt_cell_pgx_genotype)}
\usage{
rt_cell_pgx_variants(value)
}
\description{
Cell renderer for the detected variants of a gene ('; '-separated),
one variant per line (aligned with rt_cell_pgx_genotype)
}
\keyword{internal}
