% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_pdf.R
\name{typst_escape}
\alias{typst_escape}
\title{Escape Typst markup characters in plain text}
\usage{
typst_escape(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Escape Typst markup characters in plain text
}
\keyword{internal}
