% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_cpsr_output}
\alias{write_cpsr_output}
\title{Function that writes contents of CPSR report object to various output formats
(quarto HTML reports, TSV, XLSX workbooks etc)}
\usage{
write_cpsr_output(report, output_format = "html")
}
\arguments{
\item{report}{List object with all report data (CPSR), settings etc.}

\item{output_format}{file format of output
(html//tsv/xlsx etc)}
}
\description{
Function that writes contents of CPSR report object to various output formats
(quarto HTML reports, TSV, XLSX workbooks etc)
}
