% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{assign_pathogenicity_evidence}
\alias{assign_pathogenicity_evidence}
\title{Function that assigns variant pathogenicity evidence based on ACMG guidelines}
\usage{
assign_pathogenicity_evidence(var_calls, settings, ref_data)
}
\arguments{
\item{settings}{cpsr settings object}

\item{ref_data}{pcgr data object}

\item{calls}{sample calls with dbnsfp annotations}
}
\value{
calls
}
\description{
Function that assigns variant pathogenicity evidence based on ACMG guidelines
}
