% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curated_transcripts}
\alias{curated_transcripts}
\title{Curated transcripts}
\format{
A data frame with curated transcripts for MEN1
}
\usage{
curated_transcripts
}
\description{
Curated transcripts
}
\keyword{datasets}
