% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{known_path_benign_sites}
\alias{known_path_benign_sites}
\title{Function that extracts known benign peptide changes
from ClinVar (>= 2 gold stars), as well as known
pathogenic nucleotide and codon sites}
\usage{
known_path_benign_sites(ref_data = NULL)
}
\arguments{
\item{ref_data}{reference data object with ClinVar
annotations}
}
\value{
list with data.frames for benign peptide changes,
pathogenic nucleotide sites, and pathogenic codon sites
}
\description{
Function that extracts known benign peptide changes
from ClinVar (>= 2 gold stars), as well as known
pathogenic nucleotide and codon sites
}
