% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_pgx_calls}
\alias{retrieve_pgx_calls}
\title{Function that retrieves variants in genes recommended for secondary
findings}
\usage{
retrieve_pgx_calls(calls)
}
\arguments{
\item{calls}{data frame with all calls found}
}
\description{
Function that retrieves variants in genes recommended for secondary
findings
}
